Fixed whitespace and other similar errors
