+++ lib/fetchJSON.js
@@ -1,7 +1,7 @@
 var fs = require('fs');
 var Q = require('q');
 
-/** 
+/**
  * Reads from the configuration file and returns data
  * @param {string} confFile - the path and filename to the configuration file
  * @return {Q} - Q promise object
