+++ lib/sla.js
@@ -124,39 +124,35 @@ function addChannelsToBridge(client, channel, dialed, bridge, extension) {
    * @param {Object} object - contains the bridge and channel in question
    */
   function channelLeft(event, object) {
-    isStation(extension, object.channel).then(function (isStation) {
-      if (isStation) {
-
-        if (object.bridge.channels.length === 1) {
-          var hangupOutsider = Q.denodeify(client.channels.hangup.bind
-              (client.channels));
-          hangupOutsider({channelId: object.bridge.channels[0]})
-            .catch(function (err) {
-              err.name = 'HangupFailure';
-              defer.reject(err);
-            });
-        }
-      } else {
-
-        if (object.bridge.channels.length) {
-          console.log('Hanging up station channel/s');
-          var hangupArray = [];
-          object.bridge.channels.forEach( function(ID) {
-            var hangup = Q.denodeify(client.channels.hangup.bind
-              (client.channels));
-            hangupArray.push(hangup({channelId: ID}));
+    if (isStation(extension, object.channel)) {
+      if (object.bridge.channels.length === 1) {
+        var hangupOutsider = Q.denodeify(client.channels.hangup.bind
+            (client.channels));
+        hangupOutsider({channelId: object.bridge.channels[0]})
+          .catch(function (err) {
+            err.name = 'HangupFailure';
+            defer.reject(err);
           });
-          return Q.all(hangupArray);
-        }
       }
-
-      if (object.bridge.channels.length === 0) {
-        defer.resolve();
-        updateState(client, bridge.name, idle);
-        bridge.removeAllListeners('ChannelEnteredBridge');
-        bridge.removeAllListeners('ChannelLeftBridge');
+    } else {
+      if (object.bridge.channels.length) {
+        console.log('Hanging up station channel/s');
+        var hangupArray = [];
+        object.bridge.channels.forEach( function(ID) {
+          var hangup = Q.denodeify(client.channels.hangup.bind
+            (client.channels));
+          hangupArray.push(hangup({channelId: ID}));
+        });
+        return Q.all(hangupArray);
       }
-    });
+    }
+    
+    if (object.bridge.channels.length === 0) {
+      defer.resolve();
+      updateState(client, bridge.name, idle);
+      bridge.removeAllListeners('ChannelEnteredBridge');
+      bridge.removeAllListeners('ChannelLeftBridge');
+    }
   }
 }
 
@@ -192,34 +188,29 @@ function originateInbound(client, inbound, bridge, extension) {
 
     if (currentState !== inUse && currentState !== busy &&
         currentState !== ringing) {
-      extension.getAllStations()
-        .then(function (stations) {
-
-          if (stations.length) {
-            stations.forEach(function(station) {
-              dialed.push({endpoint: station, channel: client.Channel()});
-            });
-
-            dialed.forEach(function (object) {
-              var originate = Q.denodeify(object.channel.originate.bind(
-                  object.channel));
-              originate({endpoint: object.endpoint, app: 'sla',
-                  appArgs: 'dialed', timeout: 10})
-                .catch(function (err) {
-                  defer.reject(err);
-                });
-
-              updateState(client, bridge.name, ringing);
-
-              object.channel.once('StasisStart', dialedEnteredStasis);
-              object.channel.once('ChannelDestroyed', dialedChannelHungup);
-            });
-           inbound.once('ChannelHangupRequest', inboundHungup);
-          } else {
-            defer.reject(new customError.CustomError('NoStations',
-                  'No stations in this shared extension.'));
-          }
+      var stations = extension.getAllStations();
+      if (stations.length) {
+        stations.forEach(function(station) {
+          dialed.push({endpoint: station, channel: client.Channel()});
         });
+        
+        dialed.forEach(function (object) {
+          var originate = Q.denodeify(object.channel.originate.bind(
+              object.channel));
+          originate({endpoint: object.endpoint, app: 'sla',
+            appArgs: 'dialed', timeout: 10})
+            .catch(function (err) {
+              defer.reject(err);
+            });
+          updateState(client, bridge.name, ringing);
+          object.channel.once('StasisStart', dialedEnteredStasis);
+          object.channel.once('ChannelDestroyed', dialedChannelHungup);
+        });
+        inbound.once('ChannelHangupRequest', inboundHungup);
+      } else {
+        defer.reject(new customError.CustomError('NoStations',
+              'No stations in this shared extension.'));
+      }
     } else {
       // This will either redirect this channel towards a Hangup or a "backup
       // extension"
@@ -466,30 +457,29 @@ function originateOutbound(client, trunkData, outbound, bridge) {
  * @return {Q} - Q promise object
  */
 function isStation (extension, channel) {
-  return extension.getAllStations().then(function(stations) {
-    var isStation = false;
-    for (var index = 0; index < stations.length; index++) {
-      var station = stations[index];
-      var re = regexp().find(station).toRegExp();
+  var stations = extension.getAllStations()
+  var isStation = false;
+  for (var index = 0; index < stations.length; index++) {
+    var station = stations[index];
+    var re = regexp().find(station).toRegExp();
 
-      // First conditional is for regular SIP users like SIP/phone1
-      if (re.test(channel.name)) {
-        isStation = true;
-        break;
-      } else {
+    // First conditional is for regular SIP users like SIP/phone1
+    if (re.test(channel.name)) {
+      isStation = true;
+      break;
+    } else {
 
-        // Second conditional is for other SIP users that have a different
-        // server like SIP/phone2@1234
-        var nameDivided = channel.name.split('/');
-        if (re.test(nameDivided[0] + '/' + channel.caller.number + '@' +
-            nameDivided[1])) {
-              isStation = true;
-              break;
-            }
-      }
+      // Second conditional is for other SIP users that have a different
+      // server like SIP/phone2@1234
+      var nameDivided = channel.name.split('/');
+      if (re.test(nameDivided[0] + '/' + channel.caller.number + '@' +
+          nameDivided[1])) {
+            isStation = true;
+            break;
+          }
     }
-    return isStation;
-  });
+  }
+  return isStation;
 }
 
 
@@ -564,7 +554,7 @@ module.exports = function(client, confFile, channel, extensionName) {
                                  data.extension);
     })
     .then(function() {
-        return 'Extension'+ extensionName + 'freed!';
+        return 'Extension '+ extensionName + ' freed!';
     })
     .catch(function (err) {
       if (err.name !== 'InboundHungup' && err.name !== 'OutboundHungup' &&
