+++ tests/sla_tests.js
@@ -96,7 +96,7 @@ var getMockClient = function() {
           return deviceState.deviceName === params.deviceName;
         });
         if (!exists) {
-          ds = {deviceName: params.deviceName, 
+          ds = {deviceName: params.deviceName,
             deviceState: params.deviceState, hasBeenInUse: false};
           if (ds.deviceState === 'NOT_INUSE') {
             ds.isIdle = true;
@@ -140,7 +140,7 @@ var getMockClient = function() {
           cb(null, currentState);
         }
       }
-      
+
     };
 
     this.Playback = function() {
@@ -155,7 +155,7 @@ var getMockClient = function() {
 
 /**
  * Returns a mock bridge that has a bridge_type and id, and can add channels
- * to its array of channels. Emits a ChannelEnteredBridge event for every 
+ * to its array of channels. Emits a ChannelEnteredBridge event for every
  * channel that enters it.
  */
 
@@ -238,7 +238,7 @@ var getMockChannel = function() {
             if (channel === self.id) {
               channelFoundInBridges = true;
             }
-            return channel !== self.id;  
+            return channel !== self.id;
           });
 
           if (channelFoundInBridges) {
@@ -316,7 +316,7 @@ describe('SLA Bridge and Channels Tester', function() {
           bridgeChecking();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should use a preexisting bridge if there is one', function(done) {
@@ -338,7 +338,7 @@ describe('SLA Bridge and Channels Tester', function() {
           bridgeChecking();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should use a "valid" endpoint and give off a StasisStart event ' +
@@ -361,7 +361,7 @@ describe('SLA Bridge and Channels Tester', function() {
           validChecking();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should dial an "invalid" endpoint and not give off a StasisStart event',
@@ -386,7 +386,7 @@ describe('SLA Bridge and Channels Tester', function() {
           invalidChecking();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should enter the application but specify an invalid SLA bridge. The ' +
@@ -412,10 +412,10 @@ describe('SLA Bridge and Channels Tester', function() {
           incorrectBridge();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
-  it('should enter the application, call the dialed channel, but hang up ' + 
+  it('should enter the application, call the dialed channel, but hang up ' +
       'before it answers (which in turn hangs up inbound caller)',
      function(done) {
     var client = getMockClient();
@@ -442,7 +442,7 @@ describe('SLA Bridge and Channels Tester', function() {
           earlyHangup();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should hangup inbound channel if all dialed channles fail to answer',
@@ -470,7 +470,7 @@ describe('SLA Bridge and Channels Tester', function() {
       setTimeout(function() {
         if (isMixing && channels.length === 0 && bridges.length === 1 &&
           channel.inbound && channel.wasAnswered && dialed[0] &&
-          dialed[1] && dialed[0].wasHungup && dialed[1].wasHungup && 
+          dialed[1] && dialed[0].wasHungup && dialed[1].wasHungup &&
           channel.wasHungup) {
                 done();
           } else {
@@ -508,7 +508,7 @@ describe('SLA Bridge and Channels Tester', function() {
             cancelDialing();
           }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should give the application an invalid configuration file and promptly ' +
@@ -614,7 +614,7 @@ describe('SLA Bridge and Channels Tester', function() {
             markAsRinging();
           }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should mark a device as NOT_INUSE when no outbound channels answer',
@@ -655,7 +655,7 @@ describe('SLA Bridge and Channels Tester', function() {
             markAsRinging();
           }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should not mark a device INUSE when a trunk enters the shared extension',
@@ -690,7 +690,7 @@ describe('SLA Bridge and Channels Tester', function() {
         }
       }, asyncDelay);
     }
-  }); 
+  });
 
   it('should test the configuration when there are no trunks to use and ' +
       'fail out', function(done) {
@@ -739,7 +739,7 @@ describe('SLA Bridge and Channels Tester', function() {
           outboundNominal();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should test whether or not the application will fail when an incorrect' +
@@ -772,7 +772,7 @@ describe('SLA Bridge and Channels Tester', function() {
           outboundBadExtension();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should test whether or not an additional station can join in a call' +
@@ -820,7 +820,7 @@ describe('SLA Bridge and Channels Tester', function() {
           secondStationJoinsBridge();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should test whether or not an additional inbound caller gets kicked out' +
@@ -886,7 +886,7 @@ describe('SLA Bridge and Channels Tester', function() {
       setTimeout(function() {
         if (channels.length === 0 && bridgeChannels.length === 0 &&
           channel.outbound && channel.wasAnswered && channel.wasHungup &&
-          secondChannel.outbound && secondChannel.wasAnswered && 
+          secondChannel.outbound && secondChannel.wasAnswered &&
           secondChannel.wasHungup && dialed[0].wasAnswered &&
           dialed[0].nonStation && dialed[0].wasHungup) {
           done();
@@ -907,7 +907,7 @@ describe('SLA Bridge and Channels Tester', function() {
           bothStationsHangup();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should test whether the application keeps the non-station channel with ' +
@@ -944,7 +944,7 @@ describe('SLA Bridge and Channels Tester', function() {
       setTimeout(function() {
         if (channels.length === 2 && bridgeChannels.length === 2 &&
           channel.outbound && channel.wasAnswered && channel.wasHungup &&
-          secondChannel.outbound && secondChannel.wasAnswered && 
+          secondChannel.outbound && secondChannel.wasAnswered &&
           !secondChannel.wasHungup && dialed[0].wasAnswered &&
           dialed[0].nonStation && !dialed[0].wasHungup) {
           done();
@@ -964,7 +964,7 @@ describe('SLA Bridge and Channels Tester', function() {
           oneStationHangsup();
         }
       }, asyncDelay);
-    } 
+    }
   });
 
   it('should hangup both station channels when the outside channel hangs up',
@@ -1000,7 +1000,7 @@ describe('SLA Bridge and Channels Tester', function() {
       setTimeout(function() {
         if (channels.length === 0 && bridgeChannels.length === 0 &&
           channel.outbound && channel.wasAnswered && channel.wasHungup &&
-          secondChannel.outbound && secondChannel.wasAnswered && 
+          secondChannel.outbound && secondChannel.wasAnswered &&
           secondChannel.wasHungup && dialed[0].wasAnswered &&
           dialed[0].nonStation && dialed[0].wasHungup) {
           done();
